+++ stories/Toggler.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { injectGlobal } from 'styled-components';
 
 interface State {
   show: boolean;
@@ -12,6 +13,21 @@ interface ChildrenProps {
 interface Props {
   children: (props: ChildrenProps) => React.ReactNode;
 }
+
+injectGlobal`
+  button {
+    padding: 10px;
+    margin: 10px;
+    border: 2px solid black;
+    text-transform: uppercase;
+
+    &:focus {
+      border: 2px solid blue;
+      outline: 2px solid blue;
+      outline-offset: 2px;
+    }
+  }
+`;
 
 export default class Toggler extends React.Component<Props, State> {
   state = {